+++ docs/changelog.md
@@ -445,8 +445,9 @@ Many feature gates are removed in order to make most features make always availa
 #### rand
 - move `num::prob::rand` to [base].
 - rename `Lgc16` to `Lcg16`.
+- new macro: `define_pcg!`.
 - new traits: `Rand`, `RandAlloc`, `RandStd`.
-- new types: `Pcg32`.
+- new type: `Pcg32`.
 
 ### quant
 - move to [base]: `Cycle`, `CycleCount`, `Interval`,  `Sign`.